Add mock view bindings and enhance wire set cert handling
- Introduced a new mock view bindings file to simulate various endpoints for testing purposes.
- Implemented mock functions for authentication and several endpoints including /whoami, /work-item-details, /pyro-assets, and others.
- Enhanced error handling in the wire offsets tests to cover database errors and 404 scenarios.
- Added a new data class `WireSetCertResult` to encapsulate the results of wire set cert refresh operations.
- Refactored the `WireSetCertRefresher` class to utilize the new `WireSetCertResult` class for better structure and clarity.
- Improved the handling of field mapping during the refresh process to streamline data extraction from Excel files.
